+++ providers/config.rb
@@ -1,9 +1,10 @@
 #
 # Author:: Kendrick Martin (kendrick.martin@webtrends.com)
+# Contributor:: David Dvorak (david.dvorak@webtrends.com)
 # Cookbook Name:: iis
 # Resource:: config
 #
-# Copyright:: 2011, Webtrends
+# Copyright:: 2011, Webtrends Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -24,9 +25,10 @@ include Chef::Mixin::ShellOut
 include Windows::Helper
 
 action :config do
-	Chef::Log.debug("#{appcmd} set config #{@new_resource.cfg_cmd}")
-	shell_out!("#{appcmd} set config #{@new_resource.cfg_cmd}")
-	Chef::Log.info("IIS Config command run")	
+  cmd = "#{appcmd} set config #{@new_resource.cfg_cmd}"
+  Chef::Log.debug(cmd)
+  shell_out!(cmd, :returns => @new_resource.returns)
+  Chef::Log.info("IIS Config command run")
 end
 
 private