+++ recipes/remove_default_site.rb
@@ -1,9 +1,9 @@
 #
-# Author:: Seth Chisamore (<schisamo@opscode.com>)
+# Author:: Kendrick Martin (<kendrick.martin@webtrends.com>)
 # Cookbook Name:: iis
-# Attribute:: default
+# Recipe:: remove_default_site
 #
-# Copyright:: Copyright (c) 2011 Opscode, Inc.
+# Copyright 2012, Webtrends, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,12 +18,10 @@
 # limitations under the License.
 #
 
-default['iis']['accept_eula'] = false
+iis_site 'Default Web Site' do
+	action [:stop, :delete]
+end
 
-default['iis']['home']       = "#{ENV['WINDIR']}\\System32\\inetsrv"
-default['iis']['conf_dir']   = "#{iis['home']}\\config"
-default['iis']['pubroot']    = "#{ENV['SYSTEMDRIVE']}\\inetpub"
-default['iis']['docroot']    = "#{iis['pubroot']}\\wwwroot"
-default['iis']['log_dir']    = "#{iis['pubroot']}\\logs\\LogFiles"
-default['iis']['cache_dir']  = "#{iis['pubroot']}\\temp"
-default['iis']['components'] = "IIS7"
+iis_pool 'DefaultAppPool' do
+	action [:stop , :delete]
+end